properly organize test files and update dependencies
This organizes the test files more like we do with our ruby gems
and other ruby projects. This keeps the source consistent.

This updates the dependencies to keep up with the latest stuff.
