+++ Gemfile
@@ -2,4 +2,5 @@ source "https://rubygems.org"
 
 ruby "~> 2.4"
 
-gem "assert", "~> 2.17.0"
+gem "pry", "~> 0.12.2"
+gem "assert", "~> 2.18.1"
